+++ middleware/redirect/redirect.go
@@ -5,16 +5,15 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/dusted-go/http/v2/middleware/chain"
 	"github.com/dusted-go/http/v2/request"
+	"github.com/dusted-go/http/v2/server"
 )
 
 // ForceHTTPS is a middleware which redirects http:// requests to https://
-func ForceHTTPS(enable bool, hosts ...string) chain.Intermediate {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+func ForceHTTPS(enable bool, hosts ...string) server.Middleware {
+	return server.MiddlewareFunc(
+		func(next http.Handler, w http.ResponseWriter, r *http.Request) {
 			if enable {
-
 				isMatch := false
 				for _, h := range hosts {
 					if h == r.Host {
@@ -22,7 +21,6 @@ func ForceHTTPS(enable bool, hosts ...string) chain.Intermediate {
 						break
 					}
 				}
-
 				redirect := isMatch && !request.IsHTTPS(r)
 				if redirect {
 					url := request.HTTPSURL(r)
@@ -32,41 +30,41 @@ func ForceHTTPS(enable bool, hosts ...string) chain.Intermediate {
 			}
 			next.ServeHTTP(w, r)
 		})
-	}
 }
 
 // TrailingSlash is a middleware which will redirect a matching request with
 // a trailing slash in the path to the same endpoint without a trailing slash.
-func TrailingSlash(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		path := r.URL.Path
-		redirect :=
-			strings.HasPrefix(r.Proto, "HTTP") && // Must be HTTP request
-				len(path) > 1 && // Skip if it is just the root (/) path
-				path[len(path)-1] == '/' // Must have trailing slash
+func TrailingSlash() server.Middleware {
+	return server.MiddlewareFunc(
+		func(next http.Handler, w http.ResponseWriter, r *http.Request) {
+			path := r.URL.Path
+			redirect :=
+				strings.HasPrefix(r.Proto, "HTTP") && // Must be HTTP request
+					len(path) > 1 && // Skip if it is just the root (/) path
+					path[len(path)-1] == '/' // Must have trailing slash
 
-		if redirect {
-			scheme := "http"
-			if request.IsHTTPS(r) {
-				scheme = "https"
+			if redirect {
+				scheme := "http"
+				if request.IsHTTPS(r) {
+					scheme = "https"
+				}
+				url := fmt.Sprintf("%s://%s%s", scheme, r.Host, path[:len(path)-1])
+				if r.URL.RawQuery != "" {
+					url = url + "?" + r.URL.RawQuery
+				}
+				http.Redirect(w, r, url, 301)
+				return
 			}
-			url := fmt.Sprintf("%s://%s%s", scheme, r.Host, path[:len(path)-1])
-			if r.URL.RawQuery != "" {
-				url = url + "?" + r.URL.RawQuery
-			}
-			http.Redirect(w, r, url, 301)
-			return
-		}
 
-		next.ServeHTTP(w, r)
-	})
+			next.ServeHTTP(w, r)
+		})
 }
 
 // Hosts is a middleware which redirects from one host to another.
 // (e.g. https://www.foo.bar -> https://foo.bar)
-func Hosts(hosts map[string]string, enable bool) chain.Intermediate {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+func Hosts(hosts map[string]string, enable bool) server.Middleware {
+	return server.MiddlewareFunc(
+		func(next http.Handler, w http.ResponseWriter, r *http.Request) {
 			if enable {
 				if dest, ok := hosts[r.Host]; ok {
 					r.Host = dest
@@ -77,5 +75,4 @@ func Hosts(hosts map[string]string, enable bool) chain.Intermediate {
 			}
 			next.ServeHTTP(w, r)
 		})
-	}
 }